#1519 Turn GOLooSwap into a generic LopSwap transformation.
