+++ doc/user_guide/gocean1p0.rst
@@ -907,11 +907,6 @@ the GOcean 1.0 API. For an overview of transformations in general see
    :members:
    :noindex:
 
-####
-
-.. autoclass:: psyclone.transformations.GOLoopSwapTrans
-   :members:
-   :noindex:
 
 ####
